Move application mode enum into modes submodule
